pyproject.toml
[project]
name = "arcadium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "absl-py==2.3.1",
    "accelerate==1.10.1",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "certifi==2025.8.3",
    "chardet==5.2.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "dataproperty==1.1.0",
    "datasets==3.6.0",
    "dill==0.3.8",
    "einops==0.8.1",
    "evaluate==0.4.5",
    "filelock==3.19.1",
    "fonttools==4.59.1",
    "frozenlist==1.7.0",
    "fsspec==2025.3.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "hf-xet==1.1.9",
    "huggingface-hub==0.34.4",
    "idna==3.10",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "jsonlines==4.0.0",
    "kiwisolver==1.4.9",
    "lm-eval==0.4.9.1",
    "lxml==6.0.1",
    "markupsafe==3.0.2",
    "matplotlib==3.10.5",
    "mbstrdecoder==1.1.4",
    "more-itertools==10.8.0",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "multiprocess==0.70.16",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "numexpr==2.12.0",
    "numpy==2.2.6",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-cupti-cu12==12.8.90",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.3.3.83",
    "nvidia-cufile-cu12==1.13.1.3",
    "nvidia-curand-cu12==10.3.9.90",
    "nvidia-cusolver-cu12==11.7.3.90",
    "nvidia-cusparse-cu12==12.5.8.93",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.3",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvtx-cu12==12.8.90",
    "omegaconf==2.3.0",
    "packaging==25.0",
    "pandas==2.3.2",
    "pathvalidate==3.3.1",
    "peft==0.17.1",
    "pillow==11.3.0",
    "platformdirs==4.4.0",
    "plotly>=6.7.0",
    "portalocker==3.2.0",
    "propcache==0.3.2",
    "protobuf==6.32.0",
    "psutil==7.0.0",
    "pyarrow==21.0.0",
    "pybind11==3.0.1",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "pytablewriter==1.2.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "regex==2025.9.1",
    "requests==2.32.5",
    "rouge-score==0.1.2",
    "sacrebleu==2.5.1",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.15.3",
    "sentry-sdk==2.35.1",
    "setuptools==80.9.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sqlitedict==2.1.0",
    "sympy==1.14.0",
    "tabledata==1.3.4",
    "tabulate==0.9.0",
    "tcolorpy==0.1.7",
    "threadpoolctl==3.6.0",
    "tokenizers==0.22.0",
    "torch==2.8.0",
    "torchinfo==1.8.0",
    "tqdm==4.67.1",
    "tqdm-multiprocess==0.0.11",
    "transformers==4.56.1",
    "triton==3.4.0",
    "typepy==1.3.4",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wandb==0.21.3",
    "word2number==1.1",
    "xxhash==3.5.0",
    "yarl==1.20.1",
    "zstandard==0.24.0",
]
